Switch to Spring CorsFilter
